Iterantion 3 spring new functionality

+++ src/main/java/com/example/moview/moview/controller/ReviewController.java
@@ -3,9 +3,11 @@ package com.example.moview.moview.controller;
 import com.example.moview.moview.dto.review.ReviewCreateDto;
 import com.example.moview.moview.dto.review.ReviewDto;
 import com.example.moview.moview.dto.review.ReviewUpdateDto;
+import com.example.moview.moview.mapper.ReviewMapper;
 import com.example.moview.moview.model.Review;
 import com.example.moview.moview.service.ReviewService;
-import org.modelmapper.ModelMapper;
+import com.example.moview.moview.validator.ReviewValidator;
+import com.example.moview.moview.validator.UserValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -24,30 +27,46 @@ import javax.validation.Valid;
 public class ReviewController {
 
     private final ReviewService reviewService;
-    private final ModelMapper modelMapper;
+
+    private final ReviewMapper reviewMapper;
+    private final UserValidator userValidator;
+    private final ReviewValidator reviewValidator;
 
     @Autowired
-    public ReviewController(final ReviewService reviewService, final ModelMapper modelMapper) {
+    public ReviewController(final ReviewService reviewService,
+                            final ReviewMapper reviewMapper,
+                            final UserValidator userValidator,
+                            final ReviewValidator reviewValidator) {
+
         this.reviewService = reviewService;
-        this.modelMapper = modelMapper;
+        this.reviewMapper = reviewMapper;
+        this.userValidator = userValidator;
+        this.reviewValidator = reviewValidator;
     }
 
-   @PostMapping
-    public ResponseEntity<ReviewDto> create(@RequestBody @Valid final ReviewCreateDto dto) {
-        final Review review = modelMapper.map(dto, Review.class);
-        final ReviewDto reviewDto = modelMapper.map(reviewService.create(review), ReviewDto.class);
-        return ResponseEntity.status(HttpStatus.OK).body(reviewDto);
+    @PostMapping
+    public ResponseEntity<ReviewDto> create(@RequestHeader("Authorization") final Long userId,
+                                            @RequestBody @Valid final ReviewCreateDto dto) {
+
+        userValidator.validateUserExisting(userId);
+        final Review review = reviewMapper.reviewCreateDtoToEntity(dto, userId);
+        final ReviewDto reviewDto = reviewMapper.entityToReviewDto(reviewService.create(review));
+        return ResponseEntity.status(HttpStatus.CREATED).body(reviewDto);
     }
 
     @PutMapping
-    public ResponseEntity<ReviewDto> update(@RequestBody @Valid final ReviewUpdateDto dto) {
-        final Review review = modelMapper.map(dto, Review.class);
-        final ReviewDto reviewDto = modelMapper.map(reviewService.update(review), ReviewDto.class);
-        return ResponseEntity.status(HttpStatus.OK).body(reviewDto);
+    public ResponseEntity<ReviewDto> update(@RequestHeader("Authorization") final Long userId,
+                                            @RequestBody @Valid final ReviewUpdateDto dto) {
+
+        reviewValidator.validateAuthor(userId, dto.getId());
+        final Review review = reviewMapper.reviewUpdateDtoToEntity(dto, userId);
+        final ReviewDto reviewDto = reviewMapper.entityToReviewDto(reviewService.update(review));
+        return ResponseEntity.ok(reviewDto);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public void delete(@RequestHeader("Authorization") final Long userId, @PathVariable final Long id) {
+        reviewValidator.validateAuthor(userId, id);
         reviewService.delete(id);
     }
 }
